MemeOnPoint/requirements.txt
spacy
Pillow
tweepy
schedule
bs4
